backend/llm-service/pyproject.toml
[project]
name = "llm-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.11",
    "langchain-groq>=0.3.1",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-ollama>=0.3.0",
    "langgraph>=0.3.18",
    "mcp>=1.9.3",
    "python-dotenv>=1.0.1",
    "redisvl>=0.6.0",
    "tavily-python>=0.5.1",
    "uvicorn>=0.34.0",
]

